requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2
pydantic==2.5.3